게시물 dao 도입

+++ src/main/java/com/sb/fbPhoto/controller/MpaUsrArticleController.java
@@ -25,10 +25,7 @@ public class MpaUsrArticleController {
 		if (Util.isEmpty(body)) {
 			return new ResultData("F-2", "내용을 입력해주세요");
 		}
-		int id = articleService.writeArticle(title, body);
-		Article article = articleService.getArticleById(id);
-
-		return new ResultData("S-1", id + "번 글이 작성되었습니다.", "article", article);
+		return articleService.writeArticle(title, body);
 	}
 
 	@RequestMapping("/mpaUsr/article/doModify")
@@ -44,23 +41,21 @@ public class MpaUsrArticleController {
 		if (Util.isEmpty(body)) {
 			return new ResultData("F-3", "내용을 입력해주세요");
 		}
-		boolean modified = articleService.modifyArticle(id, title, body);
-		if (modified == false) {
-			return new ResultData("F-1", id + "번 글이 존재하지 않습니다.", "id", id);
+		Article article = articleService.getArticleById(id);
+		if(article == null) {
+			return new ResultData("F-4", "존재하지 않는 게시물입니다.");
 		}
-
-		return new ResultData("S-1", id + "번 글이 수정되었습니다.", "article", articleService.getArticleById(id));
+		return articleService.modifyArticle(id, title, body);
 	}
 
 	@RequestMapping("/mpaUsr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
-		boolean deleted = articleService.deleteArticleById(id);
+	public ResultData doDelete(Integer id) {
 
-		if (deleted == false) {
-			return new ResultData("F-1", id + "번 글이 존재하지 않습니다.", "id", id);
+		if (Util.isEmpty(id)) {
+			return new ResultData("F-1", "번호를 입력해주세요");
 		}
-		return new ResultData("S-1", id + "번 글이 삭제되었습니다.", "id", id);
+			return articleService.deleteArticleById(id);
 	}
 
 	@RequestMapping("/mpaUsr/article/getArticle")
